Move some label-based slicing code to variable/
