+++ python/operations.cpp
@@ -6,9 +6,9 @@
 #include "pybind11.h"
 
 #include "scipp/dataset/dataset.h"
-#include "scipp/dataset/slice.h"
 #include "scipp/dataset/sort.h"
 #include "scipp/variable/operations.h"
+#include "scipp/variable/slice.h"
 #include "scipp/variable/sort.h"
 #include "scipp/variable/util.h"
 
@@ -137,16 +137,4 @@ void init_operations(py::module &m) {
           // Do NOT release GIL since using py::slice
           return std::tuple{dim.name(), py::slice(start, stop, 1)};
         });
-  m.def(
-      "select",
-      [](const VariableConstView &var, const VariableConstView &coord,
-         const VariableConstView &value) { return select(var, coord, value); },
-      py::call_guard<py::gil_scoped_release>());
-  m.def(
-      "select",
-      [](const VariableConstView &var, const VariableConstView &coord,
-         const VariableConstView &begin, const VariableConstView &end) {
-        return select(var, coord, begin, end);
-      },
-      py::call_guard<py::gil_scoped_release>());
 }
